Splash screen created
Splash screen created with 3second timer and made it to finish using finish() so that we can come back to splash screen again and app gets exited.

Also created a sign up screen activity named SignUpScreen. From splash screen we Handled to move to the signup screen. But signupscreen is empty at this point.