Completed Upload Image function while sign up
we completed Upload Image function and it will display the image on the profile while filling details. we have also created a new activity with navigation tab below named homeactivity and commented the upper navigation to stop app from crashing because we have linked the SignUpScreen activity to this homeactivity. Also created a login activity which is still empty and needs to be linked to SignUpScreen.